+++ includes/blocks.php
@@ -330,7 +330,8 @@ function radio_station_block_editor_assets() {
 
 	// --- enqueue admin script for localized variables ---
 	$script_url = plugins_url( '/js/radio-station-admin.js', RADIO_STATION_FILE );
-	$script_path = RADIO_STATION_DIR . 'js/radio-station-admin.js';
+	// 2.5.18: fix to add missing / after dir path
+	$script_path = RADIO_STATION_DIR . '/js/radio-station-admin.js';
 	$version = filemtime( $script_path );
 	wp_enqueue_script( 'radio-station-admin', $script_url, $deps, $version, true );
 	$js = radio_station_localization_script();
@@ -339,7 +340,8 @@ function radio_station_block_editor_assets() {
 
 	// --- block editor support for conditional loading ---
 	$script_url = plugins_url( '/blocks/editor.js', RADIO_STATION_FILE );
-	$script_path = RADIO_STATION_DIR . 'blocks/editor.js';
+	// 2.5.18: fix to add missing / after dir path
+	$script_path = RADIO_STATION_DIR . '/blocks/editor.js';
 	$version = filemtime( $script_path );
 	wp_enqueue_script( 'radio-blockedit-js', $script_url, $deps, $version, true );
 
@@ -351,7 +353,8 @@ function radio_station_block_editor_assets() {
 	// $deps = array( 'wp-edit-blocks' );
 	$stylesheets = array( 'shortcodes', 'schedule' ); // 'block-editor', 'blocks'
 	foreach ( $stylesheets as $stylekey ) {
-		$style_path = RADIO_STATION_DIR . 'css/rs-' . $stylekey . '.css';
+		// 2.5.18: fix to add missing / after dir path
+		$style_path = RADIO_STATION_DIR . '/css/rs-' . $stylekey . '.css';
 		$style_url = plugins_url( 'css/rs-' . $stylekey . '.css', RADIO_STATION_FILE );
 		$version = filemtime( $style_path );
 		wp_enqueue_style( 'rs-' . $stylekey, $style_url, array(), $version, 'all' );
@@ -379,7 +382,8 @@ function radio_station_block_editor_assets() {
 	// --- enqueue radio player styles ---
 	if ( array_key_exists( 'player', $callbacks ) ) {
 		$suffix = ''; // dev temp
-		$style_path = RADIO_STATION_DIR . 'player/css/radio-player' . $suffix . '.css';
+		// 2.5.18: fix to add missing / after dir path
+		$style_path = RADIO_STATION_DIR . '/player/css/radio-player' . $suffix . '.css';
 		$style_url = plugins_url( '/player/css/radio-player' . $suffix . '.css', RADIO_STATION_FILE );
 		$version = filemtime( $style_path );
 		// 2.5.18: change radio-player style handle to stream-player