+++ includes/templates.php
@@ -27,7 +27,7 @@ if ( !defined( 'ABSPATH' ) ) exit;
 // - Show Posts Adjacent Links
 // === Query Filters ===
 // - Playlist Archive Query Filter
-// - Schedule Override Filters
+// - Language Archive Query Filter
 // === Schedule Override Filters ===
 // - Add Override Template Filters
 // - Override Show Title
@@ -1120,7 +1120,8 @@ function radio_station_get_show_post_link( $output, $format, $link, $adjacent_po
 		}
 
 		// --- generate adjacent Show link ---
-		if ( isset( $show ) ) {
+		// 2.5.18: added is_array check to ignore false values
+		if ( isset( $show ) && is_array( $show ) ) {
 
 			// 2.5.0: fix to maybe get show key data
 			if ( isset( $show['show'] ) ) {
@@ -1354,6 +1355,43 @@ function radio_station_show_playlist_query( $query ) {
 				'value' => $show_id,
 			);
 			$query->set( $meta_query );
+		}
+	}
+}
+
+// -----------------------------
+// Language Archive Query Filter
+// -----------------------------
+// 2.5.18: added for main language link to post type archive
+add_filter( 'pre_get_posts', 'radio_station_show_language_query' );
+function radio_station_show_language_query( $query ) {
+
+	if ( !is_admin() && $query->is_main_query && is_post_type_archive( RADIO_STATION_SHOW_SLUG ) ) {
+
+		if ( isset( $_REQUEST['language'] ) && ( 'main' == sanitize_text_field( wp_unslash( $_REQUEST['language'] ) ) ) ) {
+
+			// --- get main language slug ---
+			$lang = get_option( 'WPLANG' );
+			if ( !$lang ) {
+				$lang = 'en_US';
+			}
+
+			// --- exclude posts with any specified language terms other than main ---
+			$query->set( 'tax_query', array(
+				'relation' => 'OR',
+				array(
+					'taxonomy' => RADIO_STATION_LANGUAGES_SLUG,
+					'field'    => 'term_id',
+					'terms'    => array(),
+					'operator' => 'NOT EXISTS',
+				),
+				array(
+					'taxonomy' => RADIO_STATION_LANGUAGES_SLUG,
+					'field'    => 'slug',
+					'terms'    => $lang,
+					'operator' => 'IN',
+				),
+			) );
 		}
 	}
 }